+++ be/src/olap/rowset/rowset_meta.h
@@ -142,19 +142,19 @@ public:
             &&  _rowset_meta_pb.has_end_version();
     }
 
-    int start_version() const {
+    int64_t start_version() const {
         return _rowset_meta_pb.start_version();
     }
 
-    void set_start_version(int start_version) {
+    void set_start_version(int64_t start_version) {
         _rowset_meta_pb.set_start_version(start_version);
     }
     
-    int end_version() const {
+    int64_t end_version() const {
         return _rowset_meta_pb.end_version();
     }
 
-    void set_end_version(int end_version) {
+    void set_end_version(int64_t end_version) {
         _rowset_meta_pb.set_end_version(end_version);
     }
     
@@ -174,27 +174,27 @@ public:
         _rowset_meta_pb.set_num_rows(num_rows);
     }
 
-    int total_disk_size() {
+    size_t total_disk_size() {
         return _rowset_meta_pb.total_disk_size();
     }
 
-    void set_total_disk_size(int total_disk_size) {
+    void set_total_disk_size(size_t total_disk_size) {
         _rowset_meta_pb.set_total_disk_size(total_disk_size);
     }
 
-    int data_disk_size() {
+    size_t data_disk_size() {
         return _rowset_meta_pb.data_disk_size();
     }
 
-    void set_data_disk_size(int data_disk_size) {
+    void set_data_disk_size(size_t data_disk_size) {
         _rowset_meta_pb.set_data_disk_size(data_disk_size);
     }
 
-    int index_disk_size() {
+    size_t index_disk_size() {
         return _rowset_meta_pb.index_disk_size();
     }
 
-    void set_index_disk_size(int index_disk_size) {
+    void set_index_disk_size(size_t index_disk_size) {
         _rowset_meta_pb.set_index_disk_size(index_disk_size);
     }
